refactor: Move command and result classes to dto package for improved organization
