+++ delivery-service/src/main/java/com/winner/client/deliveryservice/common/event/DeliveryCreateEvent.java
@@ -1,6 +1,6 @@
 package com.winner.client.deliveryservice.common.event;
 
-import com.winner.client.deliveryservice.deliverymanagerhub.application.commnad.AssignEventCommand;
+import com.winner.client.deliveryservice.deliverymanagerhub.application.dto.commnad.AssignEventCommand;
 import java.util.UUID;
 
 public record DeliveryCreateEvent(
