+++ delivery-service/src/main/java/com/winner/client/deliveryservice/deliverymanagerhub/application/message/DeliveryDeliveryManagerHubUsecase.java
@@ -1,6 +1,6 @@
 package com.winner.client.deliveryservice.deliverymanagerhub.application.message;
 
-import com.winner.client.deliveryservice.deliverymanagerhub.application.commnad.AssignEventCommand;
+import com.winner.client.deliveryservice.deliverymanagerhub.application.dto.commnad.AssignEventCommand;
 
 
 public interface DeliveryDeliveryManagerHubUsecase {
