+++ delivery-service/src/main/java/com/winner/client/deliveryservice/deliverymanagerhub/presentation/requests/DeliveryManagerHubRegistrationRequest.java
@@ -1,6 +1,6 @@
 package com.winner.client.deliveryservice.deliverymanagerhub.presentation.requests;
 
-import com.winner.client.deliveryservice.deliverymanagerhub.application.commnad.DeliveryManagerHubRegistrationCommand;
+import com.winner.client.deliveryservice.deliverymanagerhub.application.dto.commnad.DeliveryManagerHubRegistrationCommand;
 import java.util.UUID;
 
 public record DeliveryManagerHubRegistrationRequest(
