+++ delivery-service/src/main/java/com/winner/client/deliveryservice/deliverymanagerhub/presentation/responses/DeliveryManagerHubInfo.java
@@ -1,6 +1,6 @@
 package com.winner.client.deliveryservice.deliverymanagerhub.presentation.responses;
 
-import com.winner.client.deliveryservice.deliverymanagerhub.application.result.DeliveryManagerHubInfoResult;
+import com.winner.client.deliveryservice.deliverymanagerhub.application.dto.result.DeliveryManagerHubInfoResult;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.Builder;
